brand 엔티티 수정 및 top10 querydsl 쿼리문 구현 1. koreanName과 englishName 칼럼 생성 2. 관련 서비스 및 레포지토리 로직 수정 3. top10 쿼리문 개선 -> querydsl 적용

+++ src/main/java/com/ward/ward_server/api/item/dto/BrandInfoResponse.java
@@ -6,8 +6,8 @@ public record BrandInfoResponse(
         String brandLogoImage,
         String brandKoreanName,
         String brandEnglishName,
-        Integer brandViewCount,
-        Integer brandWishCount,
+        long brandViewCount,
+        long brandWishCount,
         List<BrandItemResponse> itemList
 ) {
 }
